+++ GuaranteedRate.Sextant/Logging/Elasticsearch/ElasticsearchLogAppender.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Net;
 using System.Reflection;
 using Belikov.GenuineChannels;
 using GuaranteedRate.Sextant.Config;
@@ -18,7 +19,7 @@ namespace GuaranteedRate.Sextant.Logging.Elasticsearch
         private ConnectionSettings _settings;
         private string _indexName;
         private ElasticClient _client;
-        private static ISet<string> _tags;
+        protected static ISet<string> _tags;
         public bool AllEnabled { get; set; }
         public bool DebugEnabled { get; private set; }
         public bool InfoEnabled { get; private set; }
@@ -40,7 +41,7 @@ namespace GuaranteedRate.Sextant.Logging.Elasticsearch
         public static string ELASTICSEARCH_FATAL = "ElasticsearchLogAppender.Fatal.Enabled";
         public static string ELASTICSEARCH_TAGS = "ElasticsearchLogAppender.Tags";
         public static string ELASTICSEARCH_INDEX_NAME = "ElasticsearchLogAppender.IndexName";
-
+        public static string ELASTICSEARCH_LOG_RECURSIVELY = "ElasticsearchLogAppender.LogRecursively";
         #endregion
 
         public ElasticsearchLogAppender(IEncompassConfig config)
@@ -64,7 +65,8 @@ namespace GuaranteedRate.Sextant.Logging.Elasticsearch
             var infoEnabled = config.GetValue(ELASTICSEARCH_INFO, true);
             var debugEnabled = config.GetValue(ELASTICSEARCH_DEBUG, true);
             var fatalEnabled = config.GetValue(ELASTICSEARCH_FATAL, true);
-
+            LogRecurisively = config.GetValue(ELASTICSEARCH_LOG_RECURSIVELY, true);  
+            //e.g. if we fail writing a log to Elasticsearch, log the error to Elasticsearch.
             AllEnabled = allEnabled;
             ErrorEnabled = allEnabled || errorEnabled;
             WarnEnabled = allEnabled || warnEnabled;
@@ -113,52 +115,14 @@ namespace GuaranteedRate.Sextant.Logging.Elasticsearch
             }
         }
 
-        private  ElasticLogEvent CreateLogEvent(object data)
-        {
 
-            var fields = data as IDictionary<string, string>;
-
-            if (fields == null) return null;
-
-            var loggerName = "undefined";
-            if (fields.ContainsKey("loggerName"))
-            {
-                loggerName = fields["loggerName"];
-                loggerName = loggerName.ToLower();
-            }
-
-            //fields["timestamp"] = GetEpochTime().ToString();
-
-            fields["tags"] = JsonConvert.SerializeObject(_tags);
-
-            //tolower all the keys
-            var lcfields = new Dictionary<string, string>();
-            foreach (var key in fields.Keys)
-            {
-                if (!lcfields.ContainsKey(key.ToLower()))
-                {
-                    lcfields.Add(key.ToLower(), fields[key]);  //if you have a key like "Message" and another like "message" you get only the contents of "message."  Don't do that.
-                }
-            }
-
-            var logEvent = new ElasticLogEvent
-            {
-                loggerName =  lcfields.ContainsKey("loggername") ? lcfields["loggername"] : Assembly.GetExecutingAssembly().GetName().Name,
-                hostname = lcfields.ContainsKey("hostname") ? lcfields["hostname"] : System.Environment.MachineName,
-                timestamp = lcfields.ContainsKey("timestamp") ? DateTime.Parse(lcfields["timestamp"]) : DateTime.UtcNow,
-                level = fields.ContainsKey(Logger.LEVEL) ? fields[Logger.LEVEL] : "INFO",
-                message = lcfields.ContainsKey("message") ? lcfields["message"] : "no message",
-                process = lcfields.ContainsKey("process") ? lcfields["process"] : Assembly.GetExecutingAssembly().GetName().Name
-            };
-            return logEvent;
-        }
 
         protected override bool PostEvent(object data)
         {
             try
             {
 
-                var le = CreateLogEvent(data);
+                var le = SimpleLogEvent.Create(data, _tags);
                 if (le != null)
                 {
                     var response = _client.Index(le,