+++ GuaranteedRate.Sextant/Logging/Loggly/LogglyLogAppender.cs
@@ -6,7 +6,7 @@ using GuaranteedRate.Sextant.WebClients;
 
 namespace GuaranteedRate.Sextant.Logging.Loggly
 {
-    public class LogglyLogAppender : AsyncEventReporter, ILogAppender
+    public class LogglyLogAppender : AsyncWebEventReporter, ILogAppender
     {
         private readonly string _url = "";
         private readonly string _apiKey = "";
@@ -34,6 +34,7 @@ namespace GuaranteedRate.Sextant.Logging.Loggly
         public static string LOGGLY_DEBUG = "LogglyLogAppender.Debug.Enabled";
         public static string LOGGLY_FATAL = "LogglyLogAppender.Fatal.Enabled";
         public static string LOGGLY_TAGS = "LogglyLogAppender.Tags";
+        public static string LOGGLY_LOG_RECURSIVELY = "LogglyLogAppender.LogRecursively";
 
         #endregion